address PR comments: move api into protocol dir, small rename
