+++ protocol/block/api_test.go
@@ -2,8 +2,8 @@ package block_test
 
 import (
 	"context"
+	bapi "github.com/filecoin-project/go-filecoin/protocol/block"
 	"github.com/filecoin-project/go-filecoin/protocol/storage"
-	bapi "github.com/filecoin-project/go-filecoin/protocolapi/block"
 	ast "gx/ipfs/QmPVkJMTeRC6iBByPWdrRkD3BE5UXsj5HPzb4kPqL186mS/testify/assert"
 	req "gx/ipfs/QmPVkJMTeRC6iBByPWdrRkD3BE5UXsj5HPzb4kPqL186mS/testify/require"
 	"testing"
